+++ packages/expo-ui/android/src/main/java/expo/modules/ui/ExpoUIModule.kt
@@ -2,6 +2,7 @@ package expo.modules.ui
 
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
+import expo.modules.ui.button.Button
 
 class ExpoUIModule : Module() {
   override fun definition() = ModuleDefinition {
@@ -14,6 +15,10 @@ class ExpoUIModule : Module() {
 
     View(SwitchView::class) {
       Events("onCheckedChanged")
+    }
+
+    View(Button::class) {
+      Events("onButtonPressed")
     }
   }
 }